+++ src/App.tsx
@@ -14,7 +14,9 @@ import PlayerContainer from './PlayerContainer'
 import InfoSection from './InfoSection'
 import { Route, BrowserRouter } from 'react-router-dom'
 import { useState } from 'react';
-import CompanionAdList from './CompanionAdList';
+import CompanionAdList from './components/CompanionAdList';
+import CompanionAdContextProvider from './context/CompanionAdContextProvider';
+import CompanionAd from './components/CompanionAd';
 
 export const theme = createTheme({
   palette: {
@@ -35,7 +37,7 @@ const useStyles = makeStyles(() => ({
   },
 }))
 
-function App () {
+function App() {
   const classes = useStyles()
   const [tab, setTab] = useState(0)
 
@@ -50,34 +52,37 @@ function App () {
           <Route exact path="/">
             <DefaultErrorContextProvider>
               <AdTrackingPlaybackSessionProvider>
-                <Container>
-                  <Grid container spacing={3}>
-                    <Grid item xs={12}>
-                      <Paper square className={classes.paper}>
-                        <InfoSection/>
-                      </Paper>
+                <CompanionAdContextProvider>
+                  <Container>
+                    <Grid container spacing={3}>
+                      <Grid item xs={12}>
+                        <Paper square className={classes.paper}>
+                          <InfoSection />
+                        </Paper>
+                      </Grid>
+                      <Grid item xs={12} sm={6}>
+                        <Paper square className={classes.paper}>
+                          <PlayerContainer />
+                        </Paper>
+                      </Grid>
+                      <Grid item xs={12} sm={6}>
+                        <CompanionAd adSlotId="1" />
+                        <Paper square className={classes.paper}>
+                          <Tabs
+                            value={tab}
+                            onChange={handleTabChange}
+                            indicatorColor="primary"
+                            textColor="primary">
+                            <Tab label="Tracking Events" />
+                            <Tab label="Companion Ads" />
+                          </Tabs>
+                          {tab === 0 && <AdPodList />}
+                          {tab === 1 && <CompanionAdList />}
+                        </Paper>
+                      </Grid>
                     </Grid>
-                    <Grid item xs={12} sm={6}>
-                      <Paper square className={classes.paper}>
-                        <PlayerContainer/>
-                      </Paper>
-                    </Grid>
-                    <Grid item xs={12} sm={6}>
-                      <Paper square className={classes.paper}>
-                        <Tabs
-                          value={tab}
-                          onChange={handleTabChange}
-                          indicatorColor="primary"
-                          textColor="primary">
-                          <Tab label="Tracking Events"/>
-                          <Tab label="Companion Ads"/>
-                        </Tabs>
-                        {tab === 0 && <AdPodList/>}
-                        {tab === 1 && <CompanionAdList/>}
-                      </Paper>
-                    </Grid>
-                  </Grid>
-                </Container>
+                  </Container>
+                </CompanionAdContextProvider>
               </AdTrackingPlaybackSessionProvider>
             </DefaultErrorContextProvider>
           </Route>
