+++ src/context/CompanionAdContext.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { ICompanionAdContext } from "../../types/CompanionAds";
+
+const CompanionAdContext = createContext<ICompanionAdContext | undefined>(undefined);
+
+export default CompanionAdContext;
